Convert App to TypeScript
