+++ index.html
@@ -1,48 +1,48 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;500;700&display=swap"
-      rel="stylesheet"
-    />
-    <link rel="icon" type="image/svg+xml" href="favicon.svg" />
-    <link rel="apple-touch-icon" href="apple-touch-icon.png" sizes="180x180" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Pomodorox - Pomodoro App</title>
-    <meta
-      name="description"
-      content="Hourglass concept Pomodoro timer to boots your productivity."
-    />
-    <meta name="robots" content="index,follow" />
-    <meta name="theme-color" content="#E25F5F" />
-    <meta name="author" content="Anil Seervi" />
-    <meta name="twitter:card" content="summary_large_image" />
-    <meta name="twitter:creator" content="@linASeervi" />
-    <meta name="twitter:creator:id" content="2496313362" />
-    <meta name="twitter:site:id" content="2496313362" />
+	<head>
+		<meta charset="UTF-8" />
+		<link rel="preconnect" href="https://fonts.googleapis.com" />
+		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+		<link
+			href="https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;500;700&display=swap"
+			rel="stylesheet"
+		/>
+		<link rel="icon" type="image/svg+xml" href="favicon.svg" />
+		<link rel="apple-touch-icon" href="apple-touch-icon.png" sizes="180x180" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>Pomodorox - Pomodoro App</title>
+		<meta
+			name="description"
+			content="Hourglass concept Pomodoro timer to boots your productivity."
+		/>
+		<meta name="robots" content="index,follow" />
+		<meta name="theme-color" content="#E25F5F" />
+		<meta name="author" content="Anil Seervi" />
+		<meta name="twitter:card" content="summary_large_image" />
+		<meta name="twitter:creator" content="@linASeervi" />
+		<meta name="twitter:creator:id" content="2496313362" />
+		<meta name="twitter:site:id" content="2496313362" />
 
-    <meta
-      property="og:description"
-      content="Hourglass concept Pomodoro timer to boots your productivity."
-    />
-    <meta
-      property="og:image"
-      content="https://pomodorox.pages.dev/assets/og-thumb.png"
-    />
-    <meta property="og:image:alt" content="Pomodorox" />
-    <meta property="og:image:height" content="1080" />
-    <meta property="og:image:type" content="image/png" />
-    <meta property="og:image:width" content="1920" />
-    <meta property="og:site_name" content="Pomodorox - Pomodoro App" />
-    <meta property="og:title" content="Pomodorox - Pomodoro App" />
-    <meta property="og:type" content="website" />
-    <meta property="og:url" content="https://pomodorox.pages.dev/" />
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
-  </body>
+		<meta
+			property="og:description"
+			content="Hourglass concept Pomodoro timer to boots your productivity."
+		/>
+		<meta
+			property="og:image"
+			content="https://pomodorox.pages.dev/assets/og-thumb.png"
+		/>
+		<meta property="og:image:alt" content="Pomodorox" />
+		<meta property="og:image:height" content="1080" />
+		<meta property="og:image:type" content="image/png" />
+		<meta property="og:image:width" content="1920" />
+		<meta property="og:site_name" content="Pomodorox - Pomodoro App" />
+		<meta property="og:title" content="Pomodorox - Pomodoro App" />
+		<meta property="og:type" content="website" />
+		<meta property="og:url" content="https://pomodorox.pages.dev/" />
+	</head>
+	<body>
+		<div id="root"></div>
+		<script type="module" src="/src/main.tsx"></script>
+	</body>
 </html>
