+++ src/main.tsx
@@ -1,0 +1,13 @@
+import React from "react"
+import ReactDOM from "react-dom/client"
+import App from "./App"
+import AppContext from "./Hooks/Context"
+import "./index.css"
+
+ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
+	<React.StrictMode>
+		<AppContext>
+			<App />
+		</AppContext>
+	</React.StrictMode>
+)
